Reorganize the sources.
