+++ Sources/FlattenOperators.swift
