+++ Sources/Property.swift
@@ -68,320 +68,6 @@ public protocol ComposableMutablePropertyProtocol: MutablePropertyProtocol {
 	func modify<Result>(_ action: (inout Value) throws -> Result) rethrows -> Result
 }
 
-// Property operators.
-//
-// A composed property is a transformed view of its sources, and does not
-// own its lifetime. Its producer and signal are bound to the lifetime of
-// its sources.
-
-extension PropertyProtocol {
-	/// Lifts a unary SignalProducer operator to operate upon PropertyProtocol instead.
-	fileprivate func lift<U>(_ transform: @escaping (SignalProducer<Value, NoError>) -> SignalProducer<U, NoError>) -> Property<U> {
-		return Property(self, transform: transform)
-	}
-
-	/// Lifts a binary SignalProducer operator to operate upon PropertyProtocol instead.
-	fileprivate func lift<P: PropertyProtocol, U>(_ transform: @escaping (SignalProducer<Value, NoError>) -> (SignalProducer<P.Value, NoError>) -> SignalProducer<U, NoError>) -> (P) -> Property<U> {
-		return { otherProperty in
-			return Property(self, otherProperty, transform: transform)
-		}
-	}
-
-	/// Maps the current value and all subsequent values to a new property.
-	///
-	/// - parameters:
-	///   - transform: A closure that will map the current `value` of this
-	///                `Property` to a new value.
-	///
-	/// - returns: A property that holds a mapped value from `self`.
-	public func map<U>(_ transform: @escaping (Value) -> U) -> Property<U> {
-		return lift { $0.map(transform) }
-	}
-
-	/// Combines the current value and the subsequent values of two `Property`s in
-	/// the manner described by `Signal.combineLatest(with:)`.
-	///
-	/// - parameters:
-	///   - other: A property to combine `self`'s value with.
-	///
-	/// - returns: A property that holds a tuple containing values of `self` and
-	///            the given property.
-	public func combineLatest<P: PropertyProtocol>(with other: P) -> Property<(Value, P.Value)> {
-		return Property.combineLatest(self, other)
-	}
-
-	/// Zips the current value and the subsequent values of two `Property`s in
-	/// the manner described by `Signal.zipWith`.
-	///
-	/// - parameters:
-	///   - other: A property to zip `self`'s value with.
-	///
-	/// - returns: A property that holds a tuple containing values of `self` and
-	///            the given property.
-	public func zip<P: PropertyProtocol>(with other: P) -> Property<(Value, P.Value)> {
-		return Property.zip(self, other)
-	}
-
-	/// Forward events from `self` with history: values of the returned property
-	/// are a tuple whose first member is the previous value and whose second
-	/// member is the current value. `initial` is supplied as the first member
-	/// when `self` sends its first value.
-	///
-	/// - parameters:
-	///   - initial: A value that will be combined with the first value sent by
-	///              `self`.
-	///
-	/// - returns: A property that holds tuples that contain previous and
-	///            current values of `self`.
-	public func combinePrevious(_ initial: Value) -> Property<(Value, Value)> {
-		return lift { $0.combinePrevious(initial) }
-	}
-
-	/// Forward only values from `self` that are not considered equivalent to its
-	/// consecutive predecessor.
-	///
-	/// - note: The first value is always forwarded.
-	///
-	/// - parameters:
-	///   - isEquivalent: A closure to determine whether two values are equivalent.
-	///
-	/// - returns: A property which conditionally forwards values from `self`.
-	public func skipRepeats(_ isEquivalent: @escaping (Value, Value) -> Bool) -> Property<Value> {
-		return lift { $0.skipRepeats(isEquivalent) }
-	}
-}
-
-extension PropertyProtocol where Value: Equatable {
-	/// Forward only values from `self` that are not equal to its consecutive predecessor.
-	///
-	/// - note: The first value is always forwarded.
-	///
-	/// - returns: A property which conditionally forwards values from `self`.
-	public func skipRepeats() -> Property<Value> {
-		return lift { $0.skipRepeats() }
-	}
-}
-
-extension PropertyProtocol where Value: PropertyProtocol {
-	/// Flattens the inner property held by `self` (into a single property of
-	/// values), according to the semantics of the given strategy.
-	///
-	/// - parameters:
-	///   - strategy: The preferred flatten strategy.
-	///
-	/// - returns: A property that sends the values of its inner properties.
-	public func flatten(_ strategy: FlattenStrategy) -> Property<Value.Value> {
-		return lift { $0.flatMap(strategy) { $0.producer } }
-	}
-}
-
-extension PropertyProtocol {
-	/// Maps each property from `self` to a new property, then flattens the
-	/// resulting properties (into a single property), according to the
-	/// semantics of the given strategy.
-	///
-	/// - parameters:
-	///   - strategy: The preferred flatten strategy.
-	///   - transform: The transform to be applied on `self` before flattening.
-	///
-	/// - returns: A property that sends the values of its inner properties.
-	public func flatMap<P: PropertyProtocol>(_ strategy: FlattenStrategy, _ transform: @escaping (Value) -> P) -> Property<P.Value> {
-		return lift { $0.flatMap(strategy) { transform($0).producer } }
-	}
-
-	/// Forward only those values from `self` that have unique identities across
-	/// the set of all values that have been held.
-	///
-	/// - note: This causes the identities to be retained to check for 
-	///         uniqueness.
-	///
-	/// - parameters:
-	///   - transform: A closure that accepts a value and returns identity
-	///                value.
-	///
-	/// - returns: A property that sends unique values during its lifetime.
-	public func uniqueValues<Identity: Hashable>(_ transform: @escaping (Value) -> Identity) -> Property<Value> {
-		return lift { $0.uniqueValues(transform) }
-	}
-}
-
-extension PropertyProtocol where Value: Hashable {
-	/// Forwards only those values from `self` that are unique across the set of
-	/// all values that have been seen.
-	///
-	/// - note: This causes the identities to be retained to check for uniqueness.
-	///         Providing a function that returns a unique value for each sent
-	///         value can help you reduce the memory footprint.
-	///
-	/// - returns: A property that sends unique values during its lifetime.
-	public func uniqueValues() -> Property<Value> {
-		return lift { $0.uniqueValues() }
-	}
-}
-
-extension PropertyProtocol {
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol>(_ a: A, _ b: B) -> Property<(A.Value, B.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol>(_ a: A, _ b: B, _ c: C) -> Property<(A.Value, B.Value, C.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D) -> Property<(A.Value, B.Value, C.Value, D.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer, e.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer, e.producer, f.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol, H: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G, _ h: H) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value, H.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer, h.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol, H: PropertyProtocol, I: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G, _ h: H, _ i: I) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value, H.Value, I.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer, h.producer, i.producer) }
-	}
-
-	/// Combines the values of all the given properties, in the manner described
-	/// by `combineLatest(with:)`.
-	public static func combineLatest<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol, H: PropertyProtocol, I: PropertyProtocol, J: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G, _ h: H, _ i: I, _ j: J) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value, H.Value, I.Value, J.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.combineLatest($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer, h.producer, i.producer, j.producer) }
-	}
-
-	/// Combines the values of all the given producers, in the manner described by
-	/// `combineLatest(with:)`. Returns nil if the sequence is empty.
-	public static func combineLatest<S: Sequence>(_ properties: S) -> Property<[S.Iterator.Element.Value]>? where S.Iterator.Element: PropertyProtocol {
-		let producers = properties.map { $0.producer }
-		guard !producers.isEmpty else {
-			return nil
-		}
-
-		return Property(unsafeProducer: SignalProducer.combineLatest(producers))
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol>(_ a: A, _ b: B) -> Property<(A.Value, B.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol>(_ a: A, _ b: B, _ c: C) -> Property<(A.Value, B.Value, C.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D) -> Property<(A.Value, B.Value, C.Value, D.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer, e.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer, e.producer, f.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol, H: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G, _ h: H) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value, H.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer, h.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol, H: PropertyProtocol, I: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G, _ h: H, _ i: I) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value, H.Value, I.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer, h.producer, i.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`.
-	public static func zip<A: PropertyProtocol, B: PropertyProtocol, C: PropertyProtocol, D: PropertyProtocol, E: PropertyProtocol, F: PropertyProtocol, G: PropertyProtocol, H: PropertyProtocol, I: PropertyProtocol, J: PropertyProtocol>(_ a: A, _ b: B, _ c: C, _ d: D, _ e: E, _ f: F, _ g: G, _ h: H, _ i: I, _ j: J) -> Property<(A.Value, B.Value, C.Value, D.Value, E.Value, F.Value, G.Value, H.Value, I.Value, J.Value)> where Value == A.Value {
-		return a.lift { SignalProducer.zip($0, b.producer, c.producer, d.producer, e.producer, f.producer, g.producer, h.producer, i.producer, j.producer) }
-	}
-
-	/// Zips the values of all the given properties, in the manner described by
-	/// `zip(with:)`. Returns nil if the sequence is empty.
-	public static func zip<S: Sequence>(_ properties: S) -> Property<[S.Iterator.Element.Value]>? where S.Iterator.Element: PropertyProtocol {
-		let producers = properties.map { $0.producer }
-		guard !producers.isEmpty else {
-			return nil
-		}
-
-		return Property(unsafeProducer: SignalProducer.zip(producers))
-	}
-}
-
-extension PropertyProtocol where Value == Bool {
-	/// Create a property that computes a logical NOT in the latest values of `self`.
-	///
-	/// - returns: A property that contains the logial NOT results.
-	public func negate() -> Property<Value> {
-		return self.lift { $0.negate() }
-	}
-	
-	/// Create a property that computes a logical AND between the latest values of `self`
-	/// and `property`.
-	///
-	/// - parameters:
-	///   - property: Property to be combined with `self`.
-	///
-	/// - returns: A property that contains the logial AND results.
-	public func and(_ property: Property<Value>) -> Property<Value> {
-		return self.lift(SignalProducer.and)(property)
-	}
-	
-	/// Create a property that computes a logical OR between the latest values of `self`
-	/// and `property`.
-	///
-	/// - parameters:
-	///   - property: Property to be combined with `self`.
-	///
-	/// - returns: A property that contains the logial OR results.
-	public func or(_ property: Property<Value>) -> Property<Value> {
-		return self.lift(SignalProducer.or)(property)
-	}
-}
-
 /// A read-only property that can be observed for its changes over time. There
 /// are three categories of read-only properties:
 ///
@@ -497,32 +183,6 @@ public final class Property<Value>: PropertyProtocol {
 		self.init(initial: initial, then: SignalProducer(values))
 	}
 
-	/// Initialize a composed property by applying the unary `SignalProducer`
-	/// transform on `property`.
-	///
-	/// - parameters:
-	///   - property: The source property.
-	///   - transform: A unary `SignalProducer` transform to be applied on
-	///     `property`.
-	fileprivate convenience init<P: PropertyProtocol>(
-		_ property: P,
-		transform: @escaping (SignalProducer<P.Value, NoError>) -> SignalProducer<Value, NoError>
-	) {
-		self.init(unsafeProducer: transform(property.producer))
-	}
-
-	/// Initialize a composed property by applying the binary `SignalProducer`
-	/// transform on `firstProperty` and `secondProperty`.
-	///
-	/// - parameters:
-	///   - firstProperty: The first source property.
-	///   - secondProperty: The first source property.
-	///   - transform: A binary `SignalProducer` transform to be applied on
-	///             `firstProperty` and `secondProperty`.
-	fileprivate convenience init<P1: PropertyProtocol, P2: PropertyProtocol>(_ firstProperty: P1, _ secondProperty: P2, transform: @escaping (SignalProducer<P1.Value, NoError>) -> (SignalProducer<P2.Value, NoError>) -> SignalProducer<Value, NoError>) {
-		self.init(unsafeProducer: transform(firstProperty.producer)(secondProperty.producer))
-	}
-
 	/// Initialize a composed property from a producer that promises to send
 	/// at least one value synchronously in its start handler before sending any
 	/// subsequent event.
@@ -535,7 +195,7 @@ public final class Property<Value>: PropertyProtocol {
 	///
 	/// - parameters:
 	///   - unsafeProducer: The composed producer for creating the property.
-	fileprivate init(unsafeProducer: SignalProducer<Value, NoError>) {
+	internal init(unsafeProducer: SignalProducer<Value, NoError>) {
 		// Share a replayed producer with `self.producer` and `self.signal` so
 		// they see a consistent view of the `self.value`.
 		// https://github.com/ReactiveCocoa/ReactiveCocoa/pull/3042
